collector/requirements.txt
requests==2.31.0
supabase==2.3.0
python-dotenv==1.0.0
httpx==0.27.2